Add Product

+++ server/controllers/category.controller.js
@@ -1,5 +1,6 @@
 const createError = require('http-errors');
 const Category = require('../models/Category');
+const Product = require('../models/Product');
 
 module.exports.createCategory = async (req, res, next) => {
   try {
@@ -27,7 +28,7 @@ module.exports.getAllCategories = async (req, res, next) => {
 module.exports.getCategoryById = async (req, res, next) => {
   try {
     const { idCategory } = req.params;
-    const category = await Category.findById(idCategory);
+    const category = await Category.findById(idCategory).populate('products');
     if (!category) {
       throw createError(404, 'Category not found');
     }
@@ -48,8 +49,10 @@ module.exports.updateCategoryById = async (req, res, next) => {
     }
     res.status(200).send({ data: category });
   } catch (error) {
-    if(error.code === 11000){
-      return next(createError(409, 'Category with this name is already exists'));
+    if (error.code === 11000) {
+      return next(
+        createError(409, 'Category with this name is already exists')
+      );
     }
     next(error);
   }
@@ -58,6 +61,10 @@ module.exports.updateCategoryById = async (req, res, next) => {
 module.exports.deleteCategoryById = async (req, res, next) => {
   try {
     const { idCategory } = req.params;
+    const products = await Product.find({ category: idCategory });
+    if (products.length) {
+      throw createError(409, 'Category have products');
+    }
     const category = await Category.findByIdAndDelete(idCategory);
     if (!category) {
       throw createError(404, 'Category not found');
